fix: move approve/reject estimate buttons to more menu
