+++ packages/webapp/src/containers/Drawers/EstimateDetailDrawer/EstimateDetailActionsBar.tsx
@@ -42,7 +42,7 @@ function EstimateDetailActionsBar({
   closeDrawer,
 }) {
   // Estimate details drawer context.
-  const { estimateId, estimate } = useEstimateDetailDrawerContext();
+  const { estimateId } = useEstimateDetailDrawerContext();
 
   // History.
   const history = useHistory();
@@ -52,17 +52,6 @@ function EstimateDetailActionsBar({
     history.push(`/estimates/${estimateId}/edit`);
     closeDrawer(DRAWERS.ESTIMATE_DETAILS);
   };
-
-  // Handle cancel/confirm estimate approve.
-  const handleApproveEstimate = () => {
-    openAlert('estimate-Approve', { estimateId });
-  };
-
-  // Handle cancel/confirm estimate reject.
-  const handleRejectEstimate = () => {
-    openAlert('estimate-reject', { estimateId });
-  };
-
   // Handle delete sale estimate.
   const handleDeleteEstimate = () => {
     openAlert('estimate-delete', { estimateId });
@@ -93,50 +82,7 @@ function EstimateDetailActionsBar({
           />
           <NavbarDivider />
         </Can>
-        <Choose>
-          <Choose.When
-            condition={estimate.is_delivered && estimate.is_rejected}
-          >
-            <Can I={SaleEstimateAction.Edit} a={AbilitySubject.Estimate}>
-              <Button
-                className={Classes.MINIMAL}
-                icon={<Icon icon="check" />}
-                text={<T id={'mark_as_approved'} />}
-                onClick={handleApproveEstimate}
-              />
-            </Can>
-          </Choose.When>
-          <Choose.When
-            condition={estimate.is_delivered && estimate.is_approved}
-          >
-            <Can I={SaleEstimateAction.Edit} a={AbilitySubject.Estimate}>
-              <Button
-                className={Classes.MINIMAL}
-                icon={<Icon icon="close-black" />}
-                text={<T id={'mark_as_rejected'} />}
-                onClick={handleRejectEstimate}
-              />
-            </Can>
-          </Choose.When>
-          <Choose.When condition={estimate.is_delivered}>
-            <Can I={SaleEstimateAction.Edit} a={AbilitySubject.Estimate}>
-              <Button
-                className={Classes.MINIMAL}
-                icon={<Icon icon="check" />}
-                text={<T id={'mark_as_approved'} />}
-                onClick={handleApproveEstimate}
-              />
-            </Can>
-            <Can I={SaleEstimateAction.Edit} a={AbilitySubject.Estimate}>
-              <Button
-                className={Classes.MINIMAL}
-                icon={<Icon icon="close-black" />}
-                text={<T id={'mark_as_rejected'} />}
-                onClick={handleRejectEstimate}
-              />
-            </Can>
-          </Choose.When>
-        </Choose>
+
         <Can I={SaleEstimateAction.View} a={AbilitySubject.Estimate}>
           <Button
             className={Classes.MINIMAL}
